+++ src/remover.rs
@@ -5,8 +5,8 @@ use log::debug;
 use std::fs::{rename, File};
 use std::io::{BufRead, BufReader, BufWriter, Write};
 
-use crate::display::TodoFile;
 use crate::errors::TodoRError;
+use crate::todo::TodoFile;
 
 pub fn remove_todo_by_index(todo_file: &mut TodoFile, ind: usize) -> Result<(), Error> {
     assert!(ind < todo_file.todos.len());